backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
boto3>=1.28.0
pdfplumber>=0.10.0

numpy==1.26.4
python-jose[cryptography]>=3.3.0
openai>=1.0.0
mangum>=0.17.0
